Fixed the plug-in page.
Fixed the plug in page for both the calendar and the forms pages. Had
to create a new plugin-contact.php specifically for the form because I
had to wrap it in a div to center it. so the forms page template will
have to be changed from the plugin template to forms template.